refactor(frontend): standardize component styles
- Align border widths and colors in Accordion, Cell, and ChartCell.
- Update ActionButtons, ExecuteButton, and Button to improve visual hierarchy.
- Refine hover and active states for buttons and inputs.
- Standardize text colors and background shades with the design system.
- Tweak layout and spacing in Pagination, Sidebar, and Tooltips for better usability.

Refs: #32
