+++ frontend/src/components/Accordion/Accordion.module.css
@@ -5,7 +5,7 @@ h3 {
 
 .accordionRoot {
   border-radius: var(--border-radius-lg);
-  border: var(--border-width-medium) solid var(--color-grey);
+  border: var(--border-width-thin) solid var(--color-grey-500);
   overflow: hidden;
 }
 
@@ -40,7 +40,7 @@ h3 {
 
 .accordionTrigger:hover,
 .accordionTrigger:hover *:not(.accordionChevron) {
-  background-color: var(--color-background-dark-1);
+  background-color: var(--color-grey-200);
 }
 
 .accordionContent {
@@ -62,7 +62,7 @@ h3 {
 }
 
 .accordionChevron {
-  color: var(--color-grey-dark-2);
+  color: var(--color-grey-700);
   transition: transform 300ms cubic-bezier(0.87, 0, 0.13, 1);
   flex-shrink: 0;
 }
@@ -92,4 +92,3 @@ h3 {
     opacity: 0;
   }
 }
-
