+++ frontend/src/components/ActionButtons/ActionButtons.module.css
@@ -1,20 +1,20 @@
 .actionButtonsContainer {
-    display: flex;
-    gap: 0.5rem;
-    align-items: center;
+  display: flex;
+  gap: 0;
+  align-items: center;
 }
 
 .actionButton {
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding: 0.1rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 0.1rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
-.button-icon {    
-    font-size: var(--button-size-normal);
-    color: var(--color-accent-dark-1);
+.button-icon {
+  font-size: var(--button-size-normal);
+  color: var(--color-primary-600);
 }
