+++ frontend/src/components/Button/Button.module.css
@@ -1,45 +1,45 @@
 .button {
-    all: unset;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-    border-radius: var(--border-radius-lg);
-    border: none;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font: inherit;
+  all: unset;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+  border-radius: var(--border-radius-lg);
+  border: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font: inherit;
 }
 
 /* Variants */
 .neutral {
-    background: transparent;
-    padding: var(--space-xs);
+  background: transparent;
+  padding: var(--space-xs);
 }
 
 .icon {
-    background: transparent;
-    padding: var(--space-2xs);
+  background: transparent;
+  padding: var(--space-2xs);
 }
 
 .ghost {
-    background: transparent;
-    padding: var(--space-2xs);
+  background: transparent;
+  padding: var(--space-2xs);
 }
 
 /* Hover states */
 .button:hover:not(:disabled) {
-    background-color: var(--color-grey-light-2);
-    box-shadow: var(--shadow-button-hover);
+  background-color: var(--color-grey-300);
+  box-shadow: var(--shadow-button-hover);
 }
 
 /* Active states */
 .button:active:not(:disabled) {
-    background-color: var(--color-grey-dark-1);
+  background-color: var(--color-grey-600);
 }
 
 /* Disabled state */
 .button:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-    color: var(--color-grey-dark-3);
+  opacity: 0.5;
+  cursor: not-allowed;
+  color: var(--color-grey-800);
 }
