+++ frontend/src/components/ExecuteButton/ExecuteButton.module.css
@@ -1,34 +1,34 @@
 .button {
-    all: unset;
-    padding: 0.5rem 0.5rem;
-    display: flex;
-    cursor: pointer;
-    transition: transform 0.1s ease-out;
+  all: unset;
+  padding: 0.5rem 0.5rem;
+  display: flex;
+  cursor: pointer;
+  transition: transform 0.1s ease-out;
 }
 
 .button > span {
-    color: var(--color-green-dark-2);
-    font-variation-settings: 'FILL' 1;
-    box-shadow: var(--shadow);
-    border-radius: var(--border-radius-sm);
-    background-color: var(--color-background);    
-    transition: all 0.2s ease-out;
+  color: var(--color-green-dark-2);
+  font-variation-settings: "FILL" 1;
+  box-shadow: var(--shadow);
+  border-radius: var(--border-radius-lg);
+  background-color: var(--color-background);
+  transition: all 0.2s ease-out;
 }
-  
+
 .button:hover > span {
-    color: var(--color-green);
-    background-color: var(--color-grey-light-3);
-    transform: translateY(-1px);
-    box-shadow: var(--shadow-hover);
+  color: var(--color-green);
+  background-color: var(--color-grey-200);
+  transform: translateY(-1px);
+  box-shadow: var(--shadow-hover);
 }
 
 .button:active > span {
-    color: var(--color-green-dark-2);
-    background-color: var(--color-grey-dark-2);
-    transform: translateY(0);
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  color: var(--color-green-dark-2);
+  background-color: var(--color-grey-700);
+  transform: translateY(0);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
 
 .button:active {
-    transform: scale(0.98);
+  transform: scale(0.98);
 }
