+++ frontend/src/components/Notebooks/Notebook/Cell/Cell.module.css
@@ -1,7 +1,7 @@
 .cellContainer {
-    display: flex;
-    flex-direction: column;
-    border-radius: var(--border-radius-lg);
-    border: var(--border-width-thick) solid var(--color-grey);
-    overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  border-radius: var(--border-radius-lg);
+  border: var(--border-width-thin) solid var(--color-grey-500);
+  overflow: hidden;
 }
