+++ frontend/src/components/Notebooks/Notebook/CellOutput/CellOutput.module.css
@@ -1,18 +1,18 @@
 .container {
-    overflow-x: auto;
-    border-top: var(--border-width-thin) solid var(--color-grey-dark-1);
-    border-radius: 0;
+  overflow-x: auto;
+  border-top: var(--border-width-thin) solid var(--color-grey-600);
+  border-radius: 0;
 }
 
 .tableHeaderCell {
-    background-color: var(--color-grey);
-    padding: var(--space-xs);
-    font-size: var(--font-size-sm);
-    border-right: var(--border-width-thin) solid var(--color-grey-dark-1);
+  background-color: var(--color-grey-500);
+  padding: var(--space-xs);
+  font-size: var(--font-size-sm);
+  border-right: var(--border-width-thin) solid var(--color-grey-600);
 }
 
 .tableDataCell {
-    padding: var(--space-xs);
-    font-size: var(--font-size-sm);
-    border-right: var(--border-width-thin) solid var(--color-grey-dark-1);
+  padding: var(--space-xs);
+  font-size: var(--font-size-sm);
+  border-right: var(--border-width-thin) solid var(--color-grey-600);
 }
