+++ frontend/src/components/Notebooks/Notebook/CellToolbar/CellToolbar.module.css
@@ -1,30 +1,30 @@
 .rightSideContainer {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: var(--space-2xs);
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: var(--space-2xs);
 }
 
 .cellNameContainer {
-    padding: 0 var(--space-xs) 0 var(--space-xs);
-    color: var(--color-grey-dark-2);
-    font-size: var(--font-size-sm);
-    border-radius: var(--border-radius-lg);
+  padding: 0 var(--space-xs) 0 var(--space-xs);
+  color: var(--color-grey-700);
+  font-size: var(--font-size-sm);
+  border-radius: var(--border-radius-lg);
 }
 
 .cellNameContainer:hover {
-    background-color: var(--color-grey-light-3);
+  background-color: var(--color-grey-200);
 }
 
 .cellNameToolTipContainer {
-    font-size: var(--font-size-xs);
-	line-height: var(--line-height-tightest);
+  font-size: var(--font-size-xs);
+  line-height: var(--line-height-tightest);
 }
 
 .cellNameToolTipHeader {
-    font-size: var(--font-size-sm);
-    font-weight: var(--font-weight-bold);
-    padding-bottom: var(--space-2xs);
+  font-size: var(--font-size-sm);
+  font-weight: var(--font-weight-bold);
+  padding-bottom: var(--space-2xs);
 }
 
 .cellNameToolTipDescription {
