+++ frontend/src/components/Notebooks/Notebook/ChartCell/ChartCell.module.css
@@ -1,13 +1,13 @@
 .container {
-    display: flex;
-    flex-direction: column;
-    border-radius: var(--border-radius-lg);
-    border: var(--border-width-thick) solid var(--color-grey);
-    overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  border-radius: var(--border-radius-lg);
+  border: var(--border-width-thin) solid var(--color-grey-500);
+  overflow: hidden;
 }
 
 .editorAndOutputContainer {
-    display: flex;
-    flex-direction: row;
-    gap: var(--space-2xs);
+  display: flex;
+  flex-direction: row;
+  gap: var(--space-2xs);
 }
