+++ frontend/src/components/Notebooks/Notebook/NotebookMenuBar/NotebookMenuBar.module.css
@@ -1,4 +1,4 @@
 .icon {
-	font-size: var(--button-size-normal);
-	color: var(--color-foreground);
+  font-size: var(--button-size-normal);
+  color: var(--color-text);
 }
