+++ frontend/src/components/SearchBar/Searchbar.module.css
@@ -1,31 +1,31 @@
 .searchbar {
-    flex: 5;
-    border-radius: var(--border-radius-lg);
-    box-shadow: var(--shadow);
-    transition: all 0.3s ease;
-    padding: 0.25rem 0.25rem 0.25rem 0.25rem;
+  flex: 5;
+  border-radius: var(--border-radius-lg);
+  box-shadow: var(--shadow);
+  transition: all 0.3s ease;
+  padding: 0.25rem 0.25rem 0.25rem 0.25rem;
 }
 
 .searchbar:hover {
-    box-shadow: var(--shadow-hover);
+  box-shadow: var(--shadow-hover);
 }
 
 .searchbarInput {
-    width: 30rem;
-    height: 2rem;
-    border: none;
-    font-size: 1rem;
-    padding: 0.5rem 0.5rem 0.5rem 0.5rem;
-    outline: none;
-    transition: all 0.3s ease;
+  width: 30rem;
+  height: 2rem;
+  border: none;
+  font-size: 1rem;
+  padding: 0.5rem 0.5rem 0.5rem 0.5rem;
+  outline: none;
+  transition: all 0.3s ease;
 }
 
 .searchbarInput::placeholder {
-    color: var(--color-grey-dark-1);
-    font-style: italic;
+  color: var(--color-grey-600);
+  font-style: italic;
 }
 
 .searchbar:focus-within {
-    box-shadow: var(--shadow-hover);
-    border: var(--border-input-focus);
+  box-shadow: var(--shadow-hover);
+  border: var(--border-input-focus);
 }
