+++ frontend/src/components/Sidebar/Sidebar.module.css
@@ -1,49 +1,47 @@
 .sidebar {
-    /* padding: var(--space-xs); */
-    /* margin: var(--space-xs); */
-    background-color: var(--color-background);
-    box-shadow: var(--shadow);
-    flex: 1.5;
-    max-width: 50px;
-    border-right: var(--border-width-thin) solid var(--color-grey);
+  /* padding: var(--space-xs); */
+  /* margin: var(--space-xs); */
+  background-color: var(--color-grey-200);
+  flex: 1.5;
+  max-width: 50px;
 }
 
 .navItems {
-    display: flex;
-    flex-direction: column;
-    gap: 0;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  gap: 0;
+  align-items: center;
 }
 
 .navItem {
-    padding: 0.5rem 0.5rem 0.5rem 0.5rem;
-    border-radius: var(--border-radius-full);
-    transition: var(--transition-bg-color);
+  padding: 0.5rem 0.5rem 0.5rem 0.5rem;
+  border-radius: var(--border-radius-full);
+  transition: var(--transition-bg-color);
 }
 
 /* .navItem:hover {
-    background-color: var(--color-grey);
+    background-color: var(--color-grey-500);
 } */
 
 .companyLogo {
-    padding: 1rem 1rem 1rem 1rem;
-    color: var(--color-foreground);
-    margin-bottom: 2rem;
-    margin-left: auto;
-    margin-right: auto;
-    border-radius: var(--border-radius-full);
-    transition: var(--transition-bg-color);
+  padding: 1rem 1rem 1rem 1rem;
+  color: var(--color-text);
+  margin-bottom: 2rem;
+  margin-left: auto;
+  margin-right: auto;
+  border-radius: var(--border-radius-full);
+  transition: var(--transition-bg-color);
 }
 
 .companyLogo:hover {
-    background-color: var(--color-grey);
-    transform: translateX(2px);
+  background-color: var(--color-grey-500);
+  transform: translateX(2px);
 }
 
 .link {
-    text-decoration: none;
-    color: var(--color-foreground);
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
+  text-decoration: none;
+  color: var(--color-text);
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
 }
