+++ frontend/src/components/SophoNavigationMenu/SophoTabs.module.css
@@ -26,28 +26,28 @@
   border: none;
   padding: 0.75rem 1rem;
   font-size: var(--font-size-base);
-  color: var(--color-foreground);
+  color: var(--color-text);
   cursor: pointer;
   transition: color 0.2s ease;
 }
 
 .navList button:hover {
-  color: var(--color-accent-light-1);
+  color: var(--color-primary-400);
 }
 
 .navList li.active button {
-  color: var(--color-accent-dark-1);
+  color: var(--color-primary-600);
   font-weight: 600;
 }
 
 .navList li.active::after {
-  content: '';
+  content: "";
   position: absolute;
   bottom: -1px;
   left: 0;
   width: 100%;
   height: 2px;
-  background-color: var(--color-accent-dark-1);
+  background-color: var(--color-primary-600);
 }
 
 .navButton {
@@ -57,7 +57,7 @@
 
 .content {
   background-color: var(--color-background);
-  border-radius: var(--border-radius-md);
+  border-radius: var(--border-radius-lg);
   padding: var(--space-md);
   flex: 1;
   overflow: hidden;
