+++ frontend/src/components/SophoProgress/SophoProgress.module.css
@@ -1,28 +1,28 @@
 .container {
-	display: flex;
-	align-items: center;
-	gap: var(--space-xs);
+  display: flex;
+  align-items: center;
+  gap: var(--space-xs);
 }
 .progressRoot {
-	position: relative;
-	overflow: hidden;
-	background: var(--color-grey);
-	border-radius: 99999px;
-	width: 80px;
-	height: 10px;
-	transform: translateZ(0);
+  position: relative;
+  overflow: hidden;
+  background: var(--color-grey-500);
+  border-radius: 99999px;
+  width: 80px;
+  height: 10px;
+  transform: translateZ(0);
 }
 
 .progressIndicator {
-	background-color: var(--color-accent);
-	border-radius: 99999px;
-	width: 100%;
-	height: 100%;
-	transition: transform 660ms cubic-bezier(0.65, 0, 0.35, 1);
-	transform: translateX(calc(-100% + var(--progress)));
+  background-color: var(--color-primary-500);
+  border-radius: 99999px;
+  width: 100%;
+  height: 100%;
+  transition: transform 660ms cubic-bezier(0.65, 0, 0.35, 1);
+  transform: translateX(calc(-100% + var(--progress)));
 }
 
 .counterContainer {
-	font-size: var(--font-size-xs);
-	color: var(--color-grey-dark-2);
+  font-size: var(--font-size-xs);
+  color: var(--color-grey-700);
 }
