+++ frontend/src/components/SophoSelect/SophoSelect.module.css
@@ -1,115 +1,115 @@
 .root {
-    background-color: var(--color-grey-dark-3);
-    border-radius: var(--border-radius-normal);
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
+  background-color: var(--color-grey-800);
+  border-radius: var(--border-radius-normal);
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
 }
 
 .toolbarButton {
-	all: unset;
+  all: unset;
 }
 
 .selectTrigger {
-	display: inline-flex;
-	align-items: center;
-	justify-content: space-between;
-	border-radius: var(--border-radius-md);
-	padding: 0 5px;
-	font-size: var(--font-size-xs);
-	background-color: var(--color-background);
-	color: var(--color-foreground);
-	box-shadow: var(--shadow);
-	width: 150px;
-	overflow: hidden;
-    margin: 5px 2px 5px 5px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: space-between;
+  border-radius: var(--border-radius-lg);
+  padding: 0 5px;
+  font-size: var(--font-size-xs);
+  background-color: var(--color-background);
+  color: var(--color-text);
+  box-shadow: var(--shadow);
+  width: 150px;
+  overflow: hidden;
+  margin: 5px 2px 5px 5px;
 }
 
 .selectTrigger > span:first-child {
-	flex: 1;
-	overflow: hidden;
-	white-space: nowrap;
-	text-overflow: ellipsis;
-	text-align: left;
-	min-width: 0;
+  flex: 1;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  text-align: left;
+  min-width: 0;
 }
 
 .selectTrigger:hover {
-	background-color: var(--color-grey-dark-1);
+  background-color: var(--color-grey-600);
 }
 .selectTrigger:focus {
-	box-shadow: 0 0 0 2px var(--color-accent-dark-1);
+  box-shadow: 0 0 0 2px var(--color-primary-600);
 }
 .selectTrigger[data-placeholder] {
-	color: var(--inactive-color-normal);
+  color: var(--inactive-color-normal);
 }
 
 .selectIcon {
-	color: var(--color-foreground);
+  color: var(--color-text);
 }
 
 .selectContent {
-	overflow: hidden;
-	background-color: var(--color-background);
-	border-radius: var(--border-radius-md);
-	box-shadow: var(--shadow-hover);
+  overflow: hidden;
+  background-color: var(--color-background);
+  border-radius: var(--border-radius-lg);
+  box-shadow: var(--shadow-hover);
 }
 
 .selectViewport {
-	padding: 5px;
+  padding: 5px;
 }
 
 .selectItem {
-	font-size: 13px;
-	line-height: 1;
-	color: var(--color-foreground);
-	border-radius: var(--border-radius-sm);
-	display: flex;
-	align-items: center;
-	height: 25px;
-	padding: 0 35px 0 25px;
-	position: relative;
-	user-select: none;
+  font-size: 13px;
+  line-height: 1;
+  color: var(--color-text);
+  border-radius: var(--border-radius-lg);
+  display: flex;
+  align-items: center;
+  height: 25px;
+  padding: 0 35px 0 25px;
+  position: relative;
+  user-select: none;
 }
 
 .selectItem[data-disabled] {
-	color: var(--inactive-color-normal);
-	pointer-events: none;
+  color: var(--inactive-color-normal);
+  pointer-events: none;
 }
 .selectItem[data-highlighted] {
-	outline: none;
-	background-color: var(--color-accent-dark-1);
-	color: var(--color-background);
+  outline: none;
+  background-color: var(--color-primary-600);
+  color: var(--color-background);
 }
 
 .selectLabel {
-	padding: 0 25px;
-	font-size: 12px;
-	line-height: 25px;
-	color: var(--inactive-color-normal);
+  padding: 0 25px;
+  font-size: 12px;
+  line-height: 25px;
+  color: var(--inactive-color-normal);
 }
 
 .selectSeparator {
-	height: 1px;
-	background-color: var(--color-grey-dark-3);
-	margin: 5px;
+  height: 1px;
+  background-color: var(--color-grey-800);
+  margin: 5px;
 }
 
 .selectItemIndicator {
-	position: absolute;
-	left: 0;
-	width: 25px;
-	display: inline-flex;
-	align-items: center;
-	justify-content: center;
+  position: absolute;
+  left: 0;
+  width: 25px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
 }
 
 .selectScrollButton {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	height: 25px;
-	background-color: var(--color-background);
-	color: var(--color-foreground);
-	cursor: default;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 25px;
+  background-color: var(--color-background);
+  color: var(--color-text);
+  cursor: default;
 }
