+++ frontend/src/components/SophoToolTip/SophoToolTip.module.css
@@ -1,77 +1,77 @@
 .tooltipArrow {
-	fill: var(--color-grey-dark-2);
+  fill: var(--color-grey-700);
 }
 
 .tooltipContent {
-    border-radius: var(--border-radius-lg);
-	padding: var(--space-xs);
-	font-size: var(--font-size-xs);
-	line-height: var(--line-height-tightest);
-	color: var(--color-background);
-	background-color: var(--color-grey-dark-2);
-	box-shadow: var(--shadow-lg);
-	user-select: none;
-	animation-duration: 400ms;
-	animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
-	will-change: transform, opacity;
+  border-radius: var(--border-radius-lg);
+  padding: var(--space-xs);
+  font-size: var(--font-size-xs);
+  line-height: var(--line-height-tightest);
+  color: var(--color-background);
+  background-color: var(--color-grey-700);
+  box-shadow: var(--shadow-lg);
+  user-select: none;
+  animation-duration: 400ms;
+  animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
+  will-change: transform, opacity;
 }
 
 .tooltipContent[data-state="delayed-open"][data-side="top"] {
-	animation-name: slideDownAndFade;
+  animation-name: slideDownAndFade;
 }
 
 .tooltipContent[data-state="delayed-open"][data-side="right"] {
-	animation-name: slideLeftAndFade;
+  animation-name: slideLeftAndFade;
 }
 
 .tooltipContent[data-state="delayed-open"][data-side="bottom"] {
-	animation-name: slideUpAndFade;
+  animation-name: slideUpAndFade;
 }
 
 .tooltipContent[data-state="delayed-open"][data-side="left"] {
-	animation-name: slideRightAndFade;
+  animation-name: slideRightAndFade;
 }
 
 @keyframes slideUpAndFade {
-	from {
-		opacity: 0;
-		transform: translateY(2px);
-	}
-	to {
-		opacity: 1;
-		transform: translateY(0);
-	}
+  from {
+    opacity: 0;
+    transform: translateY(2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
 
 @keyframes slideRightAndFade {
-	from {
-		opacity: 0;
-		transform: translateX(-2px);
-	}
-	to {
-		opacity: 1;
-		transform: translateX(0);
-	}
+  from {
+    opacity: 0;
+    transform: translateX(-2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0);
+  }
 }
 
 @keyframes slideDownAndFade {
-	from {
-		opacity: 0;
-		transform: translateY(-2px);
-	}
-	to {
-		opacity: 1;
-		transform: translateY(0);
-	}
+  from {
+    opacity: 0;
+    transform: translateY(-2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
 
 @keyframes slideLeftAndFade {
-	from {
-		opacity: 0;
-		transform: translateX(2px);
-	}
-	to {
-		opacity: 1;
-		transform: translateX(0);
-	}
+  from {
+    opacity: 0;
+    transform: translateX(2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0);
+  }
 }
