+++ frontend/src/components/design-system/Button/Button.module.css
@@ -1,101 +1,103 @@
 .button {
-    align-items: center;
-    align-content: center;
-    display: flex;
-    border-radius: var(--border-radius-lg);
-    transition: transform 0.15s ease-in-out;
+  align-items: center;
+  align-content: center;
+  display: flex;
+  border-radius: var(--border-radius-lg);
+  transition: transform 0.15s ease-in-out;
 }
 
 .button:active {
-    transform: scale(0.95);
+  transform: scale(0.95);
 }
 
 .sizeSm {
-    padding: var(--space-xs);
-    font-size: var(--font-size-sm);
+  padding: var(--space-xs);
+  font-size: var(--font-size-sm);
 }
 
 .sizeMd {
-    padding: var(--space-sm);
-    font-size: var(--font-size-base);
+  padding: var(--space-sm);
+  font-size: var(--font-size-base);
 }
 
 .sizeLg {
-    padding: var(--space-md);
-    font-size: var(--font-size-base);
+  padding: var(--space-md);
+  font-size: var(--font-size-base);
 }
 
 .accent {
-    background-color: var(--color-accent);
+  background-color: var(--color-primary-500);
+  box-shadow: var(--shadow-primary-inset), var(--shadow-xs);
 }
 
 .accent:hover {
-    background-color: var(--color-accent-dark-1);
+  background-color: var(--color-primary-600);
 }
 
 .accent:active {
-    background-color: var(--color-accent-light-1);
+  background-color: var(--color-primary-400);
+  box-shadow: none;
 }
 
 .green {
-    background-color: var(--color-green-dark-1);
-    color: var(--color-background);
-    font-weight: var(--font-weight-bold);
+  background-color: var(--color-green-dark-1);
+  box-shadow: var(--shadow-green-inset), var(--shadow-xs);
 }
 
 .green:hover {
-    background-color: var(--color-green-dark-2);
+  background-color: var(--color-green-dark-2);
 }
 
 .green:active {
-    background-color: var(--color-green);
+  background-color: var(--color-green);
+  box-shadow: none;
 }
 
 .red {
-    background-color: var(--color-red);
+  background-color: var(--color-red);
 }
 
 .red:hover {
-    background-color: var(--color-red-dark-1);
+  background-color: var(--color-red-dark-1);
 }
 
 .red:active {
-    background-color: var(--color-red-dark-2);
+  background-color: var(--color-red-dark-2);
 }
 
 .grey {
-    background-color: var(--color-grey);
+  background-color: var(--color-grey-500);
 }
 
 .grey:hover {
-    background-color: var(--color-grey-dark-1);
+  background-color: var(--color-grey-600);
 }
 
 .grey:active {
-    background-color: var(--color-grey-dark-2);
+  background-color: var(--color-grey-700);
 }
 
 .transparent {
-    background-color: transparent;
+  background-color: transparent;
 }
 
 .transparent:hover {
-    background-color: var(--color-grey-light-1);
+  background-color: var(--color-grey-400);
 }
 
 .transparent:active {
-    background-color: var(--color-grey-light-2);
+  background-color: var(--color-grey-300);
 }
 
 .white {
-    background-color: var(--color-background);
-    color: var(--color-foreground);
+  background-color: var(--color-background);
+  color: var(--color-text);
 }
 
 .white:hover {
-    background-color: var(--color-grey-light-2);
+  background-color: var(--color-grey-300);
 }
 
 .white:active {
-    background-color: var(--color-grey-dark-1);
+  background-color: var(--color-grey-600);
 }
