+++ frontend/src/css/dialog.module.css
@@ -1,41 +1,41 @@
 .dialogOverlay {
-    background-color: var(--color-background);
-    position: fixed;
-    inset: 0;
-    animation: overlayShow 150ms cubic-bezier(0.16, 1, 0.3, 1);
+  background-color: var(--color-background);
+  position: fixed;
+  inset: 0;
+  animation: overlayShow 150ms cubic-bezier(0.16, 1, 0.3, 1);
 }
 
 .dialogContent {
-    background-color: var(--color-background);
-    border-radius: var(--border-radius-md);
-    box-shadow: var(--shadow);
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 90vw;
-    max-width: 25rem;
-    max-height: 50rem;
-    padding: 2rem;
+  background-color: var(--color-background);
+  border-radius: var(--border-radius-lg);
+  box-shadow: var(--shadow);
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 90vw;
+  max-width: 25rem;
+  max-height: 50rem;
+  padding: 2rem;
 }
 
 .dialogTitle {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    margin-bottom: 1.5rem;
-    padding-bottom: 1rem;
-    border-bottom: var(--border-section);
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  margin-bottom: 1.5rem;
+  padding-bottom: 1rem;
+  border-bottom: var(--border-section);
 }
 
 .dialogTitleHeading {
-    font: var(--font-primary);
-    font-size: var(--font-size-2xl);
+  font: var(--font-primary);
+  font-size: var(--font-size-2xl);
 }
 
 .dialogDescription {
-    font-size: var(--font-size-base);
-    color: var(--color-grey);
+  font-size: var(--font-size-base);
+  color: var(--color-grey-500);
 }
